# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "网络设备管理系统后端API文档",
        "title": "网络设备管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/devices": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取所有设备列表，支持分页和筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "获取设备列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码(默认1)",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量(默认10)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "设备名称搜索关键词",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "设备类型筛选",
                        "name": "device_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "设备状态筛选",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/utils.PageResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "records": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/models.Device"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建新的设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "创建设备",
                "parameters": [
                    {
                        "description": "设备信息",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Device"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/devices/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据ID获取设备详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "获取设备详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Device"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新设备信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "更新设备",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "设备信息",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Device"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除指定ID的设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "删除设备",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Device": {
            "type": "object",
            "properties": {
                "config_json": {
                    "description": "设备配置(JSON格式)",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "删除时间",
                    "type": "string"
                },
                "description": {
                    "description": "设备描述",
                    "type": "string"
                },
                "device_type": {
                    "description": "设备类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.DeviceType"
                        }
                    ]
                },
                "id": {
                    "description": "设备ID",
                    "type": "integer"
                },
                "ip": {
                    "description": "IP地址",
                    "type": "string"
                },
                "location": {
                    "description": "设备位置",
                    "type": "string"
                },
                "mac": {
                    "description": "MAC地址",
                    "type": "string"
                },
                "meta_data": {
                    "description": "元数据(JSON格式)",
                    "type": "string"
                },
                "name": {
                    "description": "设备名称",
                    "type": "string"
                },
                "status": {
                    "description": "设备状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.DeviceStatus"
                        }
                    ]
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user_id": {
                    "description": "所属用户ID",
                    "type": "integer"
                },
                "vendor": {
                    "description": "设备厂商",
                    "type": "string"
                },
                "version": {
                    "description": "设备版本",
                    "type": "string"
                }
            }
        },
        "models.DeviceStatus": {
            "type": "string",
            "enum": [
                "online",
                "offline",
                "fault"
            ],
            "x-enum-comments": {
                "DeviceStatusFault": "故障",
                "DeviceStatusOffline": "离线",
                "DeviceStatusOnline": "在线"
            },
            "x-enum-varnames": [
                "DeviceStatusOnline",
                "DeviceStatusOffline",
                "DeviceStatusFault"
            ]
        },
        "models.DeviceType": {
            "type": "string",
            "enum": [
                "user_device",
                "compute_device",
                "network_device"
            ],
            "x-enum-comments": {
                "ComputeDevice": "网络设备(路由器、交换机等)",
                "NetworkDevice": "核心网网元",
                "UserDevice": "用户设备"
            },
            "x-enum-varnames": [
                "UserDevice",
                "ComputeDevice",
                "NetworkDevice"
            ]
        },
...
        "utils.PageResult": {
            "type": "object",
            "properties": {
                "current": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "records": {
                    "description": "记录列表"
                },
                "size": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                }
            }
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "message": {
                    "description": "消息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "请在此输入 'Bearer {token}' 格式的 JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}